Changing the Bill Sequence Number to Decimal Format

+++ wcms/wcms-connection/src/main/java/org/egov/wcms/transaction/model/Connection.java
@@ -215,7 +215,7 @@ public class Connection {
 
     private String plumberName;
 
-    private Long billSequenceNumber;
+    private Double billSequenceNumber;
     
     private String manualConsumerNumber;
